Move prover struct to coordinator group
